requirements.txt
alembic==1.8.0
bcrypt==3.2.2
validate-email-address==1
blinker==1.4
cffi==1.15.0
click==8.1.3
dnspython==2.2.1
email-validator==1.2.1
filelock==3.7.1
Flask==2.1.2
Flask-Bcrypt==1.0.1
Flask-Login==0.6.1
Flask-Mail==0.9.1
Flask-Migrate==3.1.0
Flask-SQLAlchemy==2.5.1
Flask-WTF==1.0.1
future==0.18.2
greenlet==1.1.2
gunicorn==20.1.0
idna==3.3
importlib-metadata==4.11.4
importlib-resources==5.7.1
itsdangerous==2.0.1
Jinja2==3.1.2
Mako==1.2.0
MarkupSafe==2.1.1
pdfkit==1.0.0
Pillow==9.1.1
psycopg2==2.9.3
psycopg2-binary==2.9.3
py3-validate-email==1.0.5
py3dns==3.2.1
pycparser==2.21
python-dotenv==0.20.0
SQLAlchemy==1.4.37
validate-email==1.3
Werkzeug==2.1.2
wkhtmltopdf==0.2
WTForms==3.0.1
zipp==3.8.0
